requirements.txt
aws-cdk-lib==2.133.0
constructs>=10.0.0,<11.0.0
nanoid==2.0.0
nanoid-dictionary==2.4.0
python-decouple==3.8